fix: moving after attacking should cancel attack
