+++ apps/server/src/traits/combat.ts
@@ -1,7 +1,8 @@
-import type { Rect } from "@mp/math";
-import type { Tile, TimesPerSecond } from "@mp/std";
-import type { TimeSpan } from "@mp/time";
-import type { ActorId } from "../modules/world/WorldState";
+import { vec_distance, type Rect } from "@mp/math";
+import { recordValues, type Tile, type TimesPerSecond } from "@mp/std";
+import type { TickEventHandler, TimeSpan } from "@mp/time";
+import type { PatchStateMachine, ReadonlyDeep } from "@mp/sync/server";
+import type { Actor, ActorId, WorldState } from "../modules/world/WorldState";
 
 export interface CombatTrait {
   hitBox: Rect<Tile>;
@@ -13,3 +14,57 @@ export interface CombatTrait {
   attackTargetId?: ActorId;
   lastAttack?: TimeSpan;
 }
+
+export function combatBehavior(
+  state: PatchStateMachine<WorldState>,
+): TickEventHandler {
+  return ({ totalTimeElapsed }) => {
+    for (const actor of recordValues(state.actors())) {
+      attemptAttack(actor, totalTimeElapsed);
+      if (actor.health <= 0 && actor.type == "npc") {
+        state.actors.remove(actor.id);
+      }
+    }
+  };
+
+  function attemptAttack(actor: ReadonlyDeep<Actor>, currentTime: TimeSpan) {
+    if (!actor.attackTargetId) {
+      return; // Not attacking
+    }
+
+    const target = state.actors()[actor.attackTargetId];
+    if (!target) {
+      state.actors.update(actor.id, { attackTargetId: undefined });
+      return; // target doesnt exist
+    }
+
+    const distance = vec_distance(actor.coords, target.coords);
+    if (distance > actor.attackRange + tileMargin) {
+      // target too far away, move closer, but don't attack yet
+      state.actors.update(actor.id, {
+        moveTarget: target.coords,
+      });
+      return;
+    }
+
+    if (actor.lastAttack) {
+      const timeSinceLastAttack = currentTime.subtract(
+        actor.lastAttack as TimeSpan,
+      );
+      if (timeSinceLastAttack.totalSeconds < actor.attackSpeed) {
+        return; // attack on cooldown
+      }
+    }
+
+    state.actors.update(target.id, {
+      health: target.health - actor.attackDamage,
+    });
+
+    state.actors.update(actor.id, {
+      path: undefined, // stop moving when attacking
+      lastAttack: currentTime,
+    });
+  }
+}
+
+const tileMargin = Math.sqrt(2); // diagonal distance between two tiles
